// src/types/ErrorType.ts
export enum ErrorType {
    BUSINESS = 'BUSINESS_RULE',
    VALIDATION = 'VALIDATION_ERROR',
    NOT_FOUND = 'NOT_FOUND',
    UNAUTHORIZED = 'UNAUTHORIZED',
    FORBIDDEN = 'FORBIDDEN',
    CONFLICT = 'CONFLICT',
    DATABASE = 'DATABASE_ERROR',
    INTERNAL = 'INTERNAL_SERVER_ERROR',
    ERROR_SDK_LIVE = 'ERROR_SDK_LIVE'
}